Added failsafe to delete family

+++ main.py
@@ -155,7 +155,7 @@ def createFamily():
                 return redirect(url_for('familyPannel', familyID = SQLFamilyID))
     return render_template('createfamily.html')
 
-@app.route('/deletefamily/')
+@app.route('/deletefamily/', methods=['GET','POST'])
 @login_required
 def deleteFamily():
     database = connect('FamilyCentral')
@@ -175,14 +175,17 @@ def deleteFamily():
         if SQLIsAdmin == 0:
             return redirect(url_for('familyPannel', familyID = SQLFamilyID))
         else:
-            print(f"[S] Family (ID: {SQLFamilyID}) was deleted")
-            cursor.execute('DELETE FROM Family WHERE FamilyID=' + str(SQLFamilyID) + ';')
-            cursor.execute('DELETE FROM Event WHERE FamilyID=' + str(SQLFamilyID) + ';')
-            cursor.execute('DELETE FROM ShoppingListItem WHERE FamilyID=' + str(SQLFamilyID) + ';')
-            cursor.execute('UPDATE User SET IsAdmin = 0 WHERE FamilyID="' + str(SQLFamilyID) + '";')
-            cursor.execute('UPDATE User SET FamilyID = NULL WHERE FamilyId="' + str(SQLFamilyID) + '";')
-            database.commit()
-        return redirect(url_for('home'))
+            if request.method == 'POST':
+                print(f"[S] Family (ID: {SQLFamilyID}) was deleted")
+                cursor.execute('DELETE FROM Family WHERE FamilyID=' + str(SQLFamilyID) + ';')
+                cursor.execute('DELETE FROM Event WHERE FamilyID=' + str(SQLFamilyID) + ';')
+                cursor.execute('DELETE FROM Invite WHERE FamilyID=' + str(SQLFamilyID) + ';')
+                cursor.execute('DELETE FROM ShoppingListItem WHERE FamilyID=' + str(SQLFamilyID) + ';')
+                cursor.execute('UPDATE User SET IsAdmin = 0 WHERE FamilyID="' + str(SQLFamilyID) + '";')
+                cursor.execute('UPDATE User SET FamilyID = NULL WHERE FamilyId="' + str(SQLFamilyID) + '";')
+                database.commit()
+                return redirect(url_for('home'))    
+            return render_template('deleteFamily.html')
 
 ##################################
 # END CREATE AND DELETE FAMILIES #
